Added List for GUI, WIP
List has items which contain a texture and a string of text, with (soon to be) their own callback function for each item. I also added some deepcopy abilities for some utility classes, made the naming convention more cohesive, and worked out the CPU bug. The bug fix is known to sometimes cause game stutters, but seeing as this is a roguelike game, a stutter every now and then is worth full usage of a CPU core